+++ src/rl_framework/util/__init__.py
@@ -18,7 +18,7 @@ from .features_extractor_utils import (
     wrap_environment_with_features_extractor_preprocessor,
 )
 from .imitation_data_loader import create_memory_efficient_transition_batcher
-from .training_callbacks import (
+from .sb3_training_callbacks import (
     LoggingCallback,
     SavingCallback,
     add_callbacks_to_callback,
